pyproject.toml
[project]
name = "freight-market-sensing"
version = "0.1.0"
description = ""
authors = [
    {name = "Timothy Ghrist",email = "tghrist@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pandas (>=3.0.2,<4.0.0)",
    "fredapi (>=0.5.2,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
